+++ src/main/java/sixgaezzang/sidepeek/users/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package sixgaezzang.sidepeek.users.dto;
+package sixgaezzang.sidepeek.users.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
